+++ CHANGELOG.md
@@ -27,6 +27,9 @@ This is an example of a brief overview of the _Major_ or _Minor_ version changes
 
 ---
 
+## [9.16] Added ZeroState component
+- Added `ZeroState` component, useful for displaying when a lack of the primary UI isn't available (due to lack of results, columns, etc)
+
 ## [9.15] Updatable dot
 - Added the `showDot` prop to `Updatable` to support showing an `Indicator` dot when the
   `message` is empty.